requirements.txt
Pillow
numpy
spidev
gpiozero
smbus2
RPi.GPIO
